CXP-1291: moved useChannelLocales to the shared namespace
