+++ components/catalogs/front/src/components/ProductMapping/components/SelectChannelLocaleDropdown.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components';
 import {Field, Locale, SelectInput} from 'akeneo-design-system';
 import {useTranslate} from '@akeneo-pim-community/shared';
 import {Source} from '../models/Source';
-import {useChannelLocales} from '../../ProductSelection/hooks/useChannelLocales';
+import {useChannelLocales} from '../../../hooks/useChannelLocales';
 
 const DropdownField = styled(Field)`
     margin-top: 10px;
