+++ components/catalogs/front/src/hooks/useChannelLocales.unit.ts
@@ -3,7 +3,7 @@ jest.unmock('./useChannelLocales');
 import {renderHook} from '@testing-library/react-hooks';
 import {useChannelLocales} from './useChannelLocales';
 import fetchMock from 'jest-fetch-mock';
-import {ReactQueryWrapper} from '../../../../tests/ReactQueryWrapper';
+import {ReactQueryWrapper} from '../../tests/ReactQueryWrapper';
 
 test('it fetches the API response', async () => {
     fetchMock.mockResponseOnce(
